Refatorar redirecionamento de páginas
