+++ src/Pages/Tutorial/index.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useHistory } from "react-router-dom";
 
 import { BackImg, ButtonDiv, CurrentDiv,
@@ -64,24 +64,6 @@ export default function Tutorial() {
             currentPage.classList.toggle('currentPage');
         pageIndex[page + direcao].classList.add('currentPage');
     }
-
-    const isIos = () => {
-        const userAgent = window.navigator.userAgent.toLowerCase();
-        return /iphone|ipad|ipod/.test( userAgent );
-    }
-
-    const isInStandaloneMode = () => { //referencia: https://stackoverflow.com/questions/21125337/how-to-detect-if-web-app-running-standalone-on-chrome-mobile
-        if(isIos())
-            return ('standalone' in window.navigator) && (window.navigator.standalone)
-        else 
-            return (window.matchMedia('(display-mode: standalone)').matches);
-    };   
-
-    useEffect(() => {
-        if(isInStandaloneMode()) { 
-            history.push('/Restaurante')
-        }
-    })
 
     return(
         <TutDiv>
